+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed defines descriptions and values.
 - Fixed semantic resolution for `enum struct` scope operator (`::`) and `methodmap` static methods.
 - Fixed `static` methodmap's method support.
+- Fixed highlighting inside array access.
 
 ## [4.0.6]